test: split server_test.go into focused files
- server_helpers_test.go: Test helper functions
- server_init_test.go: Server initialization tests
- server_client_test.go: Client connection tests
- server_broadcast_test.go: Event broadcasting tests
- server_shutdown_test.go: Shutdown tests
- server_subscription_test.go: Subscription routing tests
